Update custom_actions_template.py
Trying to align to the DataExchange block in Block.py. Please assess this as it's important.

we need slots input_data and output_data defined in the Rasa config i think and we will need to update processing in the data_exchange function, please feel free to move this to another branch but we must track it otherwise integrating Rasa will be difficult

Co-Authored-By: Bhavani915875 <121215222+Bhavani915875@users.noreply.github.com>
Co-Authored-By: lintaox <141118195+lintaox@users.noreply.github.com>